// src/orval-test/client.schemas.ts
export interface AdminDto {
  id: number;
  name: string;
  phoneNumber: string;
  createTime: string;
  gender: string;
  role: string;
}